+++ src/BusinessLogic/AdminAPI/Connection/Responses/OnboardingDataResponse.php
@@ -24,13 +24,20 @@ class OnboardingDataResponse extends Response
     protected $statisticalData;
 
     /**
+     * @var string|null
+     */
+    protected $portalUrl;
+
+    /**
      * @param ConnectionData[] $connectionData
      * @param StatisticalData|null $statisticalData
+     * @param string|null $portalUrl URL of the SeQura portal the store is connected to
      */
-    public function __construct(array $connectionData, ?StatisticalData $statisticalData)
+    public function __construct(array $connectionData, ?StatisticalData $statisticalData, ?string $portalUrl = null)
     {
         $this->connectionData = $connectionData;
         $this->statisticalData = $statisticalData;
+        $this->portalUrl = $portalUrl;
     }
 
     /**
@@ -43,6 +50,7 @@ class OnboardingDataResponse extends Response
         }
 
         $response['sendStatisticalData'] = $this->statisticalData && $this->statisticalData->isSendStatisticalData();
+        $response['portalUrl'] = $this->portalUrl;
 
         foreach ($this->connectionData as $connectionData) {
             $response['environment'] = $connectionData->getEnvironment();