+++ src/BusinessLogic/CheckoutAPI/CheckoutAPI.php
@@ -27,7 +27,7 @@ class CheckoutAPI
     /**
      * Gets an CheckoutAPI instance.
      *
-     * @return Aspects
+     * @return CheckoutAPI
      */
     public static function get(): object
     {