+++ tests/BusinessLogic/Common/BaseTestCase.php
@@ -328,7 +328,8 @@ class BaseTestCase extends TestCase
                 return new ConnectionService(
                     TestServiceRegister::getService(ConnectionDataRepositoryInterface::class),
                     TestServiceRegister::getService(CredentialsService::class),
-                    TestServiceRegister::getService(StoreIntegrationService::class)
+                    TestServiceRegister::getService(StoreIntegrationService::class),
+                    TestServiceRegister::getService(DeploymentsRepositoryInterface::class)
                 );
             },
             AffiliateSettingsRepositoryInterface::class => function () {
@@ -702,7 +703,8 @@ class BaseTestCase extends TestCase
                 return new MockConnectionService(
                     TestServiceRegister::getService(ConnectionDataRepositoryInterface::class),
                     TestServiceRegister::getService(CredentialsService::class),
-                    TestServiceRegister::getService(StoreIntegrationService::class)
+                    TestServiceRegister::getService(StoreIntegrationService::class),
+                    TestServiceRegister::getService(DeploymentsRepositoryInterface::class)
                 );
             }
         );